refactor: change location of ProfileButton component
