--- src/onboarding/components/Slider/styles.css
@@ -1,6 +1,0 @@
-.navigation-buttons{
-  position: fixed;
-  bottom: 0;
-  z-index: 1;
-  height: 50px;
-}
